requirements.txt
anyio==3.5.0
asgiref==3.5.0
asttokens==2.0.5
async-exit-stack==1.0.1
async-generator==1.10
attrs==21.4.0
autoflake==1.4
backcall==0.2.0
bcrypt==3.2.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
coverage==6.3.2
cryptography==36.0.2
decorator==5.1.1
ecdsa==0.17.0
executing==0.8.3
fastapi==0.75.2
flake8==4.0.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
ipython==8.2.0
isort==5.10.1
jedi==0.18.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy==0.942
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
passlib==1.7.4
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyflakes==2.4.0
Pygments==2.12.0
pyparsing==3.0.8
pytest==7.1.2
pytest-cov==3.0.0
python-decouple==3.6
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
stack-data==0.2.0
starlette==0.17.1
tomli==2.0.1
traitlets==5.1.1
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
wcwidth==0.2.5
Werkzeug==2.1.2
